move some tests from predictor to phases
