+++ tests/unit_tests/libs/controllers/test_predictor.py
@@ -66,215 +66,6 @@ class TestPredictor(unittest.TestCase):
         # 1 call when sampling testing data for lightwood
         assert sample_settings['sample_function'].call_count == 3
 
-    def test_analyze_dataset(self):
-        n_points = 100
-        n_category_values = 4
-        input_dataframe = pd.DataFrame({
-            'numeric_int': [x % 10 for x in list(range(n_points))],
-            'numeric_float': np.linspace(0, n_points, n_points),
-            'date_timestamp': [(datetime.now() - timedelta(minutes=int(i))).isoformat() for i in range(n_points)],
-            'date_date': [(datetime.now() - timedelta(days=i)).strftime('%Y-%m-%d') for i in range(n_points)],
-            'categorical_str': [f'a{x}' for x in (list(range(n_category_values)) * (n_points // n_category_values))],
-            'categorical_int': [x for x in (list(range(n_category_values)) * (n_points // n_category_values))],
-            'categorical_binary': [0, 1] * (n_points // 2),
-            'sequential_array': [f"1,2,3,4,5,{i}" for i in range(n_points)]
-        }, index=list(range(n_points)))
-
-        model_data = F.analyse_dataset(from_data=input_dataframe)
-        for col, col_data in model_data['data_analysis_v2'].items():
-            if col == 'columns':
-                continue
-            expected_type = test_column_types[col][0]
-            expected_subtype = test_column_types[col][1]
-            assert col_data['typing']['data_type'] == expected_type
-            assert col_data['typing']['data_subtype'] == expected_subtype
-
-            assert col_data['empty']
-            assert col_data['histogram']
-            assert 'percentage_buckets' in col_data
-            assert 'nr_warnings' in col_data
-            assert col_data['identifier'] is None
-
-        assert isinstance(json.dumps(model_data), str)
-
-    def test_sample_for_analysis(self):
-        n_points = 100
-        n_category_values = 4
-        input_dataframe = pd.DataFrame({
-            'numeric_int': [x % 10 for x in list(range(n_points))],
-            'numeric_float': np.linspace(0, n_points, n_points),
-            'date_timestamp': [(datetime.now() - timedelta(minutes=int(i))).isoformat() for i in range(n_points)],
-            'date_date': [(datetime.now() - timedelta(days=i)).strftime('%Y-%m-%d') for i in range(n_points)],
-            'categorical_str': [f'a{x}' for x in (list(range(n_category_values)) * (n_points // n_category_values))],
-            'categorical_int': [x for x in (list(range(n_category_values)) * (n_points // n_category_values))],
-            'categorical_binary': [0, 1] * (n_points // 2),
-            'sequential_array': [f"1,2,3,4,5,{i}" for i in range(n_points)]
-        }, index=list(range(n_points)))
-
-        mock_function = PickableMock(spec=sample_data, wraps=sample_data)
-        setattr(mock_function, '__name__', 'mock_sample_data')
-        with mock.patch('mindsdb_native.libs.controllers.predictor.sample_data', mock_function):
-            model_data = F.analyse_dataset(
-                from_data=input_dataframe,
-                sample_settings={'sample_for_analysis': True}
-            )
-            assert mock_function.called
-
-        for col, col_data in model_data['data_analysis_v2'].items():
-            if col == 'columns':
-                continue
-            expected_type = test_column_types[col][0]
-            expected_subtype = test_column_types[col][1]
-            assert col_data['typing']['data_type'] == expected_type
-            assert col_data['typing']['data_subtype'] == expected_subtype
-
-            assert col_data['empty']
-            assert col_data['histogram']
-            assert 'percentage_buckets' in col_data
-            assert 'nr_warnings' in col_data
-            assert col_data['identifier'] is None
-
-    def test_ignore_columns(self):
-        input_dataframe = pd.DataFrame({
-            'do_use': list(range(100)),
-            'y': list(range(100)),
-            'ignore_this': list(range(100, 0, -1))
-        })
-
-        predictor = Predictor(name='test')
-        predictor.learn(
-            from_data=input_dataframe,
-            to_predict='y',
-            ignore_columns=['ignore_this'],
-            stop_training_in_x_seconds=1,
-            use_gpu=False,
-            advanced_args={'force_column_usage': ['do_use']}
-        )
-        transaction = predictor.transaction
-
-        assert 'do_use' in transaction.input_data.train_df.columns
-        assert 'ignore_this' not in transaction.input_data.train_df.columns
-
-    def test_ignore_identifiers(self):
-        input_dataframe = pd.DataFrame({
-            'do_use': [*range(60), *range(40)],
-            'numeric_id': list(range(100)),
-            'malicious_naming': list(range(99)) + [200],
-            'y': list(range(100)),
-        })
-
-        predictor = Predictor(name='test')
-        predictor.learn(
-            from_data=input_dataframe,
-            to_predict='y',
-            stop_training_in_x_seconds=1,
-            use_gpu=False
-        )
-
-        assert 'do_use' in predictor.transaction.input_data.train_df.columns
-        # Foreign key is ignored and removed from data frames
-        assert 'numeric_id' not in predictor.transaction.input_data.train_df.columns
-        assert 'numeric_id' in predictor.transaction.lmd['columns_to_ignore']
-        assert 'malicious_naming' not in predictor.transaction.input_data.train_df.columns
-        assert 'malicious_naming' in predictor.transaction.lmd['columns_to_ignore']
-
-        predictor = Predictor(name='test')
-        predictor.learn(
-            from_data=input_dataframe,
-            to_predict='y',
-            stop_training_in_x_seconds=1,
-            advanced_args={'force_column_usage': ['numeric_id']},
-            use_gpu=False
-        )
-
-        assert 'do_use' in predictor.transaction.input_data.train_df.columns
-        assert 'numeric_id' in predictor.transaction.input_data.train_df.columns
-        assert 'numeric_id' not in predictor.transaction.lmd['columns_to_ignore']
-
-    def test_analyze_dataset_empty_column(self):
-        n_points = 100
-        input_dataframe = pd.DataFrame({
-            'numeric_int': [x % 10 for x in list(range(n_points))],
-            'empty_column': [None for i in range(n_points)]
-        }, index=list(range(n_points)))
-
-        model_data = F.analyse_dataset(from_data=input_dataframe)
-
-        assert model_data['data_analysis_v2']['empty_column']['empty']['is_empty'] is True
-
-    def test_analyze_dataset_empty_values(self):
-        n_points = 100
-        input_dataframe = pd.DataFrame({
-            'numeric_int': [x % 10 for x in list(range(n_points))],
-            'numeric_int2': list(range(n_points)),
-        }, index=list(range(n_points)))
-        input_dataframe['numeric_int'].iloc[::2] = None
-
-        model_data = F.analyse_dataset(from_data=input_dataframe)
-
-        assert model_data['data_analysis_v2']['numeric_int']['empty']['empty_percentage'] == 50
-
-    def test_predictor_deduplicate_data(self):
-        n_points = 100
-        input_dataframe = pd.DataFrame({
-            'numeric_int': [x % 44 for x in list(range(n_points))],
-            'numeric_int_2': [x % 20 for x in list(range(n_points))],
-        }, index=list(range(n_points)))
-        input_dataframe['y'] = input_dataframe['numeric_int'] % 10
-
-        # Add duplicate row
-        input_dataframe = input_dataframe.append(input_dataframe.iloc[99], ignore_index=True)
-
-        mdb = Predictor(name='test_drop_duplicates')
-        mdb.learn(
-            from_data=input_dataframe,
-            to_predict='y',
-            stop_training_in_x_seconds=1,
-            use_gpu=False
-        )
-
-        model_data = F.get_model_data('test_drop_duplicates')
-
-        # Ensure duplicate row was not used for training, or analysis
-
-        assert model_data['data_preparation']['total_row_count'] == n_points
-        assert model_data['data_preparation']['used_row_count'] <= n_points
-
-        assert sum([model_data['data_preparation']['train_row_count'],
-                    model_data['data_preparation']['validation_row_count'],
-                    model_data['data_preparation']['test_row_count']]) == n_points
-
-        assert sum([mdb.transaction.input_data.train_df.shape[0],
-                    mdb.transaction.input_data.test_df.shape[0],
-                    mdb.transaction.input_data.validation_df.shape[0]]) == n_points
-
-        # Disable deduplication and ensure the duplicate row is used
-        mdb = Predictor(name='test_drop_duplicates')
-        mdb.learn(
-            from_data=input_dataframe,
-            to_predict='y',
-            stop_training_in_x_seconds=1,
-            use_gpu=False,
-            advanced_args={
-                'deduplicate_data': False
-            }
-        )
-
-        model_data = F.get_model_data('test_drop_duplicates')
-
-        # Duplicate row was used for analysis and training
-
-        assert model_data['data_preparation']['total_row_count'] == n_points+1
-        assert model_data['data_preparation']['used_row_count'] <= n_points+1
-
-        assert sum([model_data['data_preparation']['train_row_count'],
-                    model_data['data_preparation']['validation_row_count'],
-                    model_data['data_preparation']['test_row_count']]) == n_points+1
-
-        assert sum([mdb.transaction.input_data.train_df.shape[0],
-                    mdb.transaction.input_data.test_df.shape[0],
-                    mdb.transaction.input_data.validation_df.shape[0]]) == n_points+1
-
     def test_explain_prediction(self):
         mdb = Predictor(name='test_explain_prediction')
 
@@ -661,29 +452,3 @@ class TestPredictor(unittest.TestCase):
         score = f1_score(test_tags_encoded, pred_labels_encoded, average='weighted')
 
         assert score >= 0.3
-
-    def test_user_provided_split_indices(self):
-        predictor = Predictor('test')
-
-        df = pd.DataFrame({
-            'col_a': [*range(100)],
-            'col_b': [*range(100)]
-        })
-
-        predictor.learn(
-            from_data=df,
-            to_predict='col_b',
-            advanced_args={
-                'data_split_indexes': {
-                    'validation_indexes': [*range(0, 30)],
-                    'train_indexes': [*range(30, 60)],
-                    'test_indexes': [*range(60, 100)]
-                },
-                'force_column_usage': ['col_a', 'col_b']
-            },
-            use_gpu=False
-        )
-
-        assert set(predictor.transaction.input_data.train_df['col_a'].tolist()) == set(range(30, 60))
-        assert set(predictor.transaction.input_data.test_df['col_a'].tolist()) == set(range(60, 100))
-        assert set(predictor.transaction.input_data.validation_df['col_a'].tolist()) == set(range(0, 30))
