+++ tests/unit_tests/libs/phases/data_analyzer/test_data_analyzer.py
@@ -19,7 +19,38 @@ from unit_tests.utils import (
 
 
 class TestDataAnalyzer(unittest.TestCase):
-    def test_data_analysis(self):
+    def test_data_analysis_1(self):
+        n_points = 100
+        n_category_values = 4
+        df = pd.DataFrame({
+            'numeric_int': [x % 10 for x in list(range(n_points))],
+            'numeric_float': np.linspace(0, n_points, n_points),
+            'date_timestamp': [(datetime.now() - timedelta(minutes=int(i))).isoformat() for i in range(n_points)],
+            'date_date': [(datetime.now() - timedelta(days=i)).strftime('%Y-%m-%d') for i in range(n_points)],
+            'categorical_str': [f'a{x}' for x in (list(range(n_category_values)) * (n_points // n_category_values))],
+            'categorical_int': [x for x in (list(range(n_category_values)) * (n_points // n_category_values))],
+            'categorical_binary': [0, 1] * (n_points // 2),
+            'sequential_array': [f"1,2,3,4,5,{i}" for i in range(n_points)]
+        })
+
+        model_data = analyse_dataset(from_data=df)
+        for col, col_data in model_data['data_analysis_v2'].items():
+            if col == 'columns':
+                continue
+            expected_type = test_column_types[col][0]
+            expected_subtype = test_column_types[col][1]
+            assert col_data['typing']['data_type'] == expected_type
+            assert col_data['typing']['data_subtype'] == expected_subtype
+
+            assert col_data['empty']
+            assert col_data['histogram']
+            assert 'percentage_buckets' in col_data
+            assert 'nr_warnings' in col_data
+            assert col_data['identifier'] is None
+
+        assert isinstance(json.dumps(model_data), str)
+
+    def test_data_analysis_2(self):
         n_points = 100
         n_category_values = 4
         df = pd.DataFrame({
@@ -65,7 +96,7 @@ class TestDataAnalyzer(unittest.TestCase):
     
         assert stats['numeric_int']['empty']['empty_percentage'] == 50
 
-    def test_sample_true(self):
+    def test_sample_true_1(self):
         N = 100
         df = pd.DataFrame({'numeric_int': [x % 10 for x in [*range(N)]]})
         df['numeric_int'].iloc[::2] = None
@@ -97,6 +128,43 @@ class TestDataAnalyzer(unittest.TestCase):
         assert not sample_settings['sample_function'].called
         assert sum(stats['numeric_int']['histogram']['y']) <= N
 
+    def test_sample_true_2(self):
+        n_points = 100
+        n_category_values = 4
+        input_dataframe = pd.DataFrame({
+            'numeric_int': [x % 10 for x in list(range(n_points))],
+            'numeric_float': np.linspace(0, n_points, n_points),
+            'date_timestamp': [(datetime.now() - timedelta(minutes=int(i))).isoformat() for i in range(n_points)],
+            'date_date': [(datetime.now() - timedelta(days=i)).strftime('%Y-%m-%d') for i in range(n_points)],
+            'categorical_str': [f'a{x}' for x in (list(range(n_category_values)) * (n_points // n_category_values))],
+            'categorical_int': [x for x in (list(range(n_category_values)) * (n_points // n_category_values))],
+            'categorical_binary': [0, 1] * (n_points // 2),
+            'sequential_array': [f"1,2,3,4,5,{i}" for i in range(n_points)]
+        }, index=list(range(n_points)))
+
+        mock_function = PickableMock(spec=sample_data, wraps=sample_data)
+        setattr(mock_function, '__name__', 'mock_sample_data')
+        with mock.patch('mindsdb_native.libs.controllers.predictor.sample_data', mock_function):
+            model_data = analyse_dataset(
+                from_data=input_dataframe,
+                sample_settings={'sample_for_analysis': True}
+            )
+            assert mock_function.called
+
+        for col, col_data in model_data['data_analysis_v2'].items():
+            if col == 'columns':
+                continue
+            expected_type = test_column_types[col][0]
+            expected_subtype = test_column_types[col][1]
+            assert col_data['typing']['data_type'] == expected_type
+            assert col_data['typing']['data_subtype'] == expected_subtype
+
+            assert col_data['empty']
+            assert col_data['histogram']
+            assert 'percentage_buckets' in col_data
+            assert 'nr_warnings' in col_data
+            assert col_data['identifier'] is None
+
     def test_guess_probability(self):
         df = pd.DataFrame({
             'categorical_int': [1, 2, 1, 3, 4, 3, 2, 4, 5, 1, 2, 3],
@@ -107,3 +175,26 @@ class TestDataAnalyzer(unittest.TestCase):
         stats = analyse_dataset(df)['data_analysis_v2']
 
         assert stats['categorical_binary']['guess_probability'] == (9 / 12)**2 + (3 / 12)**2
+
+    def test_analyze_dataset_empty_column(self):
+        n_points = 100
+        df = pd.DataFrame({
+            'numeric_int': [x % 10 for x in list(range(n_points))],
+            'empty_column': [None for i in range(n_points)]
+        })
+
+        model_data = analyse_dataset(from_data=df)
+
+        assert model_data['data_analysis_v2']['empty_column']['empty']['is_empty'] is True
+
+    def test_analyze_dataset_empty_values(self):
+        n_points = 100
+        df = pd.DataFrame({
+            'numeric_int': [x % 10 for x in list(range(n_points))],
+            'numeric_int2': list(range(n_points)),
+        })
+        df['numeric_int'].iloc[::2] = None
+
+        model_data = analyse_dataset(from_data=df)
+
+        assert model_data['data_analysis_v2']['numeric_int']['empty']['empty_percentage'] == 50
